app/requirements.txt
flask
psycopg2-binary
flask-smorest
python-dotenv
